refactored mol-repr/visual
