+++ src/mol-repr/structure/units-visual.ts
@@ -5,7 +5,8 @@
  */
 
 import { Unit, Structure } from 'mol-model/structure';
-import { RepresentationProps, Visual, VisualContext } from '../representation';
+import { RepresentationProps } from '../representation';
+import { Visual, VisualContext } from '../visual';
 import { StructureMeshParams, StructurePointsParams, StructureLinesParams, StructureDirectVolumeParams } from './representation';
 import { Loci, isEveryLoci, EmptyLoci } from 'mol-model/loci';
 import { MeshRenderObject, PointsRenderObject, LinesRenderObject, DirectVolumeRenderObject } from 'mol-gl/render-object';
@@ -31,7 +32,6 @@ import { SizeTheme } from 'mol-theme/size';
 import { UnitsParams } from './units-representation';
 import { RenderableState } from 'mol-gl/renderable';
 import { Mat4 } from 'mol-math/linear-algebra';
-import { setTransform } from 'mol-geo/geometry/transform-data';
 
 export type StructureGroup = { structure: Structure, group: Unit.SymmetryGroup }
 
@@ -169,6 +169,7 @@ export function UnitsVisual<P extends UnitsParams>(builder: UnitsVisualGeometryB
             }
 
             if (updateState.updateTransform || updateState.createGeometry) {
+                console.log('UnitsVisual.updateBoundingSphere')
                 updateBoundingSphere(renderObject.values, newGeometry || geometry)
             }
 
@@ -239,13 +240,13 @@ export function UnitsVisual<P extends UnitsParams>(builder: UnitsVisualGeometryB
             return changed
         },
         setVisibility(value: boolean) {
-            if (renderObject) renderObject.state.visible = value
+            Visual.setVisibility(renderObject, value)
         },
         setPickable(value: boolean) {
-            if (renderObject) renderObject.state.pickable = value
+            Visual.setPickable(renderObject, value)
         },
         setTransform(value: Mat4) {
-            if (renderObject) setTransform(value, renderObject.values)
+            Visual.setTransform(renderObject, value)
         },
         destroy() {
             // TODO
